Remove checklist container pointer events from Project View cards
To select a todo and open a Todo View, the event handler checks if the card and not a child element was clicked. The checklist container occluded the click area for this event for no functional reason, so its pointer events were removed. Its child elements (checklist items) need pointer events to eneter Edit Mode, so pointer events were manually set to all, since they inherit the parent style setting that was implemented.